refactor and move to a separate module
